LRCI-3961 Move portal specific logic into the Portal open pull event handler
