+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/github/BaseOpenGitHubPullRequestEventHandler.java
@@ -6,10 +6,8 @@
 package com.liferay.jethr0.event.github;
 
 import com.liferay.jethr0.event.EventHandlerContext;
-import com.liferay.jethr0.event.github.file.GitHubFile;
 import com.liferay.jethr0.event.github.pullrequest.GitHubPullRequest;
 import com.liferay.jethr0.event.github.user.GitHubUser;
-import com.liferay.jethr0.git.branch.GitBranchEntity;
 import com.liferay.jethr0.job.JobEntity;
 import com.liferay.jethr0.util.StringUtil;
 
@@ -36,23 +34,21 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 	public String process() throws InvalidJSONException, IOException {
 		if (checkLiferayGitHubUser() ||
 			closeInvalidUpstreamGitHubBranchName() ||
-			_skipCISenderBlacklistGitHubUser()) {
+			skipCISenderBlacklistGitHubUser()) {
 
 			return null;
 		}
 
-		_commentAutoCommentMessage();
-		_commentBroadcastMessage();
+		commentAutoCommentMessage();
+		commentBroadcastMessage();
 
-		if (_checkForwardedPullRequest() ||
-			_checkMergeSubrepositoryPullRequest()) {
-
-			_commentDefaultMessage();
+		if (checkForwardedPullRequest()) {
+			commentDefaultMessage();
 
 			return null;
 		}
 
-		_invokeJobEntities();
+		invokeJobEntities();
 
 		return String.valueOf(getMessageJSONObject());
 	}
@@ -63,7 +59,7 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 		super(eventHandlerContext, messageJSONObject);
 	}
 
-	private boolean _checkForwardedPullRequest()
+	protected boolean checkForwardedPullRequest()
 		throws InvalidJSONException, IOException {
 
 		GitHubPullRequest gitHubPullRequest = getGitHubPullRequest();
@@ -97,67 +93,7 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 		return false;
 	}
 
-	private boolean _checkMergeSubrepositoryPullRequest()
-		throws InvalidJSONException, IOException {
-
-		GitHubPullRequest gitHubPullRequest = getGitHubPullRequest();
-
-		if (gitHubPullRequest == null) {
-			return false;
-		}
-
-		GitBranchEntity upstreamGitBranchEntity = getUpstreamGitBranchEntity();
-
-		GitHubUser receiverGitHubUser =
-			gitHubPullRequest.getReceiverGitHubUser();
-
-		String subrepositoryMergeReceiverUserName =
-			getJenkinsBranchBuildPropertyValue(
-				StringUtil.combine(
-					"subrepo.merge.receiver.name[",
-					upstreamGitBranchEntity.getBranchName(), "]"));
-
-		if (!gitHubPullRequest.isMergeSubrepositoryPullRequest() ||
-			!Objects.equals(
-				receiverGitHubUser.getName(),
-				subrepositoryMergeReceiverUserName)) {
-
-			return false;
-		}
-
-		GitRepo gitRepo = new GitRepo(
-			upstreamGitBranchEntity.getFileContent(
-				gitHubPullRequest.getGitRepoFilePath()));
-
-		GitHubFile ciMergeGitHubFile = gitHubPullRequest.getCIMergeGitHubFile();
-
-		Matcher matcher = _branchSHAPattern.matcher(
-			ciMergeGitHubFile.getPatch());
-
-		String ciMergeBranchSHA = "";
-
-		if (matcher.find()) {
-			ciMergeBranchSHA = matcher.group("branchSHA");
-		}
-
-		String compareURL = StringUtil.combine(
-			"https://github.com/liferay/", gitRepo.getRepositoryName(),
-			"/compare/", gitRepo.getRepositorySHA(), "..." + ciMergeBranchSHA);
-
-		gitHubPullRequest.comment(
-			StringUtil.combine(
-				"Subrepo changes: ", compareURL, "\n\nci:test:sf and ",
-				"ci:test:relevant must pass in order for auto-merge to ",
-				"initiate."));
-
-		invokeJobEntity(createPortalPullRequestJobEntity("relevant"));
-
-		invokeJobEntity(createPortalPullRequestJobEntity("sf"));
-
-		return true;
-	}
-
-	private void _commentAutoCommentMessage()
+	protected void commentAutoCommentMessage()
 		throws InvalidJSONException, IOException {
 
 		GitHubPullRequest gitHubPullRequest = getGitHubPullRequest();
@@ -181,7 +117,7 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 				"\n"));
 	}
 
-	private void _commentBroadcastMessage()
+	protected void commentBroadcastMessage()
 		throws InvalidJSONException, IOException {
 
 		String broadcastMessage = getCIProperty(
@@ -196,7 +132,7 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 		gitHubPullRequest.comment(broadcastMessage);
 	}
 
-	private void _commentDefaultMessage() throws InvalidJSONException {
+	protected void commentDefaultMessage() throws InvalidJSONException {
 		GitHubPullRequest gitHubPullRequest = getGitHubPullRequest();
 
 		GitHubUser senderGitHubUser = gitHubPullRequest.getSenderGitHubUser();
@@ -231,19 +167,10 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 		gitHubPullRequest.comment(sb.toString());
 	}
 
-	private Set<JobEntity> _createJobEntities()
-		throws InvalidJSONException, IOException {
+	protected abstract Set<JobEntity> createJobEntities()
+		throws InvalidJSONException, IOException;
 
-		Set<JobEntity> jobEntities = new HashSet<>();
-
-		for (String testSuite : _getTestSuites()) {
-			jobEntities.add(createPortalPullRequestJobEntity(testSuite));
-		}
-
-		return jobEntities;
-	}
-
-	private Set<String> _getTestSuites()
+	protected Set<String> getTestSuites()
 		throws InvalidJSONException, IOException {
 
 		GitHubPullRequest gitHubPullRequest = getGitHubPullRequest();
@@ -290,15 +217,17 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 		return testSuites;
 	}
 
-	private void _invokeJobEntities() throws InvalidJSONException, IOException {
-		Set<JobEntity> jobEntities = _createJobEntities();
+	protected void invokeJobEntities()
+		throws InvalidJSONException, IOException {
+
+		Set<JobEntity> jobEntities = createJobEntities();
 
 		for (JobEntity jobEntity : jobEntities) {
 			invokeJobEntity(jobEntity);
 		}
 	}
 
-	private boolean _skipCISenderBlacklistGitHubUser()
+	protected boolean skipCISenderBlacklistGitHubUser()
 		throws InvalidJSONException, IOException {
 
 		GitHubPullRequest gitHubPullRequest = getGitHubPullRequest();
@@ -336,45 +265,7 @@ public abstract class BaseOpenGitHubPullRequestEventHandler
 		return true;
 	}
 
-	private static final Pattern _branchSHAPattern = Pattern.compile(
-		"\\+(?<branchSHA>[0-9a-f]{40})");
 	private static final Pattern _ciTestAutoRecipientPattern = Pattern.compile(
 		"(?<userName>[^\\]]+)\\[(?<testSuites>[^\\]]+)\\]");
-
-	private static class GitRepo {
-
-		public String getRepositoryName() {
-			Matcher matcher = _gitRepoRepositoryNamePattern.matcher(
-				_gitRepoFileContent);
-
-			if (!matcher.find()) {
-				return null;
-			}
-
-			return matcher.group("repositoryName");
-		}
-
-		public String getRepositorySHA() {
-			Matcher matcher = _gitRepoSHAPattern.matcher(_gitRepoFileContent);
-
-			if (!matcher.find()) {
-				return null;
-			}
-
-			return matcher.group("repositorySHA");
-		}
-
-		private GitRepo(String gitRepoFileContent) {
-			_gitRepoFileContent = gitRepoFileContent;
-		}
-
-		private static final Pattern _gitRepoRepositoryNamePattern =
-			Pattern.compile("remote = .*/(?<repositoryName>[^\\.]*)\\.git");
-		private static final Pattern _gitRepoSHAPattern = Pattern.compile(
-			"commit = (?<repositorySHA>[0-9a-f]{40})");
-
-		private final String _gitRepoFileContent;
-
-	}
 
 }
